refactor the order import from being a class_eval on Spree::Order from api to a dedicated importer in spree_core to be reused by other parts. update the year in the licence files, added licence file to spree_core as well since this was missing.
