+++ build.sh
@@ -35,3 +35,7 @@ cd ../frontend; set_gemfile; bundle update; bundle exec rspec spec
 # Spree Sample
 echo "Setup Spree Sample and running RSpec..."
 cd ../sample; bundle install; bundle exec rake test_app; bundle exec rspec spec
+
+# Spree Hub
+echo "Setup Spree Hub and running RSpec..."
+cd ../hub; bundle install; bundle exec rake test_app; bundle exec rspec spec
