+++ core/lib/spree/core/importer.rb
@@ -1,0 +1,8 @@
+module Spree
+  module Core
+    module Importer
+    end
+  end
+end
+
+require 'spree/core/importer/order'
