+++ lib/spree.rb
