+++ license.md
@@ -1,7 +1,7 @@
 Spree License
 =============
 
-Copyright © 2007-2013, Spree Commerce Inc. and other contributors.
+Copyright © 2007-2014, Spree Commerce Inc. and other contributors.
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
